# api-sigea/openapi/sigec.yaml
openapi: 3.0.0
info:
  title: SigecAPIDMS
  description: Rest Api che consente la gestione e il monitoraggio degli eventi della regione Puglia
  version: v1
servers:
  - url: "http://localhost:9292/SigecAPI"
    description: LOCALE
  - url: ND
    description: SVILUPPO
  - url: ND
    description: COLLAUDO
  - url: ND
    description: PRODUZIONE
paths:
...
  /progetti:
    get:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: API per la lettura dei progetti per Partita IVA
      operationId: getProgettiByPIVA
      parameters:
        - name: partitaIva
          in: query
          description: Partita IVA su cui effettuare al ricerca
          required: false
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgettoModel"
          description: Oggetto contenente i progetti per Partita IVA
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
components:
  schemas:
...
    ProgettoModel:
      type: object
      properties:
        progettoId:
          type: string
          description: Id del progetto
        titoloProgetto:
          type: string
          description: Titolo del progetto
        partitaIva:
          type: string
          description: Partita IVA associata al progetto
...
    Problem:
      properties:
        detail:
          description: "A human readable explanation specific to this occurrence of the\nproblem. You MUST NOT expose internal informations, personal\ndata or implementation details through this field.\n"
          type: string
        instance:
          description: "An absolute URI that identifies the specific occurrence of the problem.\nIt may or may not yield further information if dereferenced.\n"
          format: uri
          type: string
        status:
          description: "The HTTP status code generated by the origin server for this occurrence\nof the problem.\n"
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        title:
          description: "A short, summary of the problem type. Written in english and readable\nfor engineers (usually not suited for non technical stakeholders and\nnot localized); example: Service Unavailable\n"
          type: string
        type:
          default: "about:blank"
          description: "An absolute URI that identifies the problem type.  When dereferenced,\nit SHOULD provide human-readable documentation for the problem type\n(e.g., using HTML).\n"
          format: uri
          type: string
      type: object
  responses:
...
    400BadRequest:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Bad Request
...
    404NotFound:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Not Found
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
...
    503ServiceUnavailable:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Service Unavailable
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
    401Unauthorized:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Unauthorized
    default:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Unexpected error
  headers:
    Retry-After:
      description: "Retry contacting the endpoint *at least* after seconds.\nSee https://tools.ietf.org/html/rfc7231#section-7.1.3"
      schema:
        format: int32
        type: integer
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        format: int32
        type: integer
    X-RateLimit-Remaining:
      description: The number of remaining requests in the current period
      schema:
        format: int32
        type: integer
    X-RateLimit-Reset:
      description: The number of seconds left in the current period
      schema:
        format: int32
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer